Add files via upload
Note: I followed the videos for the text adventure, but got a little lost after lesson 6.4, and I was looking to make a more stylized version of what I had for a final product. I followed a different, but similar tutorial using buttons and constants instead of classes. Let me know if having the classes is necessary, I can rework it. Hope you enjoy though!